refactor(app, components): parametersTable relocated to Components directory for PL
closes AUTH-240
