+++ app/src/organisms/ProtocolDetails/ProtocolParameters/index.tsx
@@ -1,18 +1,16 @@
 import * as React from 'react'
 import { useTranslation } from 'react-i18next'
-import styled from 'styled-components'
 
 import {
-  BORDERS,
   DIRECTION_COLUMN,
   Flex,
   SPACING,
   StyledText,
   TYPOGRAPHY,
+  ParametersTable,
+  NoParameters,
 } from '@opentrons/components'
 import { Banner } from '../../../atoms/Banner'
-import { NoParameter } from './NoParameter'
-import { formatRunTimeParameterValue } from './utils'
 
 import type { RunTimeParameter } from '@opentrons/shared-data'
 
@@ -47,112 +45,11 @@ export function ProtocolParameters({
               </StyledText>
             </Flex>
           </Banner>
-          <ProtocolParameterItems runTimeParameters={runTimeParameters} />
+          <ParametersTable runTimeParameters={runTimeParameters} t={t} />
         </Flex>
       ) : (
-        <NoParameter />
+        <NoParameters t={t} />
       )}
     </Flex>
   )
 }
-
-interface ProtocolParameterItemsProps {
-  runTimeParameters: RunTimeParameter[]
-}
-
-function ProtocolParameterItems({
-  runTimeParameters,
-}: ProtocolParameterItemsProps): JSX.Element {
-  const { t } = useTranslation('protocol_details')
-  const formatRange = (
-    runTimeParameter: RunTimeParameter,
-    minMax: string
-  ): string => {
-    const { type } = runTimeParameter
-    const choices =
-      'choices' in runTimeParameter ? runTimeParameter.choices : []
-    const count = choices.length
-
-    switch (type) {
-      case 'int':
-      case 'float':
-        return minMax
-      case 'boolean':
-        return t('on_off')
-      case 'str':
-        if (count > 2) {
-          return t('choices', { count })
-        } else {
-          return choices.map(choice => choice.displayName).join(', ')
-        }
-    }
-    return ''
-  }
-
-  return (
-    <StyledTable>
-      <thead>
-        <StyledTableHeader>{t('name')}</StyledTableHeader>
-        <StyledTableHeader>{t('default_value')}</StyledTableHeader>
-        <StyledTableHeader>{t('range')}</StyledTableHeader>
-      </thead>
-      <tbody>
-        {runTimeParameters.map((parameter: RunTimeParameter, index: number) => {
-          const min = 'min' in parameter ? parameter.min : 0
-          const max = 'max' in parameter ? parameter.max : 0
-          return (
-            <StyledTableRow
-              isLast={index === runTimeParameters.length - 1}
-              key={`runTimeParameter-${index}`}
-            >
-              <StyledTableCell isLast={index === runTimeParameters.length - 1}>
-                <StyledText as="p">{parameter.displayName}</StyledText>
-              </StyledTableCell>
-              <StyledTableCell isLast={index === runTimeParameters.length - 1}>
-                <StyledText as="p">
-                  {formatRunTimeParameterValue(parameter, t)}
-                </StyledText>
-              </StyledTableCell>
-              <StyledTableCell isLast={index === runTimeParameters.length - 1}>
-                <StyledText as="p">
-                  {formatRange(parameter, `${min}-${max}`)}
-                </StyledText>
-              </StyledTableCell>
-            </StyledTableRow>
-          )
-        })}
-      </tbody>
-    </StyledTable>
-  )
-}
-
-const StyledTable = styled.table`
-  width: 100%;
-  border-collapse: collapse;
-  text-align: left;
-`
-
-const StyledTableHeader = styled.th`
-  ${TYPOGRAPHY.labelSemiBold}
-  padding: ${SPACING.spacing8};
-  border-bottom: ${BORDERS.lineBorder};
-`
-
-interface StyledTableRowProps {
-  isLast: boolean
-}
-
-const StyledTableRow = styled.tr<StyledTableRowProps>`
-  padding: ${SPACING.spacing8};
-  border-bottom: ${props => (props.isLast ? 'none' : BORDERS.lineBorder)};
-`
-
-interface StyledTableCellProps {
-  isLast: boolean
-}
-
-const StyledTableCell = styled.td<StyledTableCellProps>`
-  padding-left: ${SPACING.spacing8};
-  padding-top: ${SPACING.spacing12};
-  padding-bottom: ${props => (props.isLast ? 0 : SPACING.spacing12)};
-`
