+++ components/src/index.ts
@@ -22,7 +22,7 @@ export * from './tabbedNav'
 export * from './slotmap'
 export * from './structure'
 export * from './tooltips'
-
+export * from './parametersTable'
 // styles
 export * from './styles'
 // new ui-overhaul style vars
