+++ shared-data/js/helpers/index.ts
@@ -28,6 +28,7 @@ export * from './getOccludedSlotCountForModule'
 export * from './labwareInference'
 export * from './getAddressableAreasInProtocol'
 export * from './getSimplestFlexDeckConfig'
+export * from './formatRunTimeParameterValue'
 
 export const getLabwareDefIsStandard = (def: LabwareDefinition2): boolean =>
   def?.namespace === OPENTRONS_LABWARE_NAMESPACE
